🚚 Move DirectionSet to separate file
